requirements.txt
biopython
pandas
numpy
cutadapt